+++ pyqgisserver/handlers/__init__.py
@@ -1,5 +1,5 @@
 
 from .basehandler import BaseHandler, NotFoundHandler # noqa F401
 from .statushandler import StatusHandler, PingHandler # noqa F401
-from .owshandler  import OwsHandler, OwsFilterHandler, OwsApiHandler # noqa F401
+from .owshandler  import OwsHandler, OwsFilterHandler, OwsApiHandler, OwsApiFilterHandler # noqa F401