+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Handle `QGIS_PROJECT_FILE` environment variable
+    - Return 400 on invalid OWS service request (no Map)
+
 ## 1.7.6 - 2021-09-28
 
 * Fix wrong type for AMQP Logger variable `connection_delay`